refactor: 히스토리 선택 컴포넌트 분리
